+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Transforms and enum initializer from [`ember-flexberry`](https://github.com/Flexberry/ember-flexberry) addon.
 
 ### Changed
+- Function `enumCaptions` now returns object without null values.
 - Offline support:
     - It is not necessary now to specify projection when syncing down or reading data from offline storage.